+++ src/env/browser/providers.ts
@@ -1,11 +1,11 @@
-import { Container } from '../../container';
-import { GitCommandOptions } from '../../git/commandOptions';
+import type { Container } from '../../container';
+import type { GitCommandOptions } from '../../git/commandOptions';
 // Force import of GitHub since dynamic imports are not supported in the WebWorker ExtensionHost
+import type { GitProvider } from '../../git/gitProvider';
+import type { IntegrationAuthenticationService } from '../../plus/integrations/authentication/integrationAuthentication';
 import { GitHubGitProvider } from '../../plus/integrations/providers/github/githubGitProvider';
-import { GitProvider } from '../../git/gitProvider';
 import { RepositoryWebPathMappingProvider } from './pathMapping/repositoryWebPathMappingProvider';
 import { WorkspacesWebPathMappingProvider } from './pathMapping/workspacesWebPathMappingProvider';
-import { IntegrationAuthenticationService } from '../../plus/integrations/authentication/integrationAuthentication';
 
 export function git(_options: GitCommandOptions, ..._args: any[]): Promise<string | Buffer> {
 	return Promise.resolve('');
@@ -21,6 +21,7 @@ export function gitLogStreamTo(
 	return Promise.resolve([[''], 0]);
 }
 
+// eslint-disable-next-line @typescript-eslint/require-await
 export async function getSupportedGitProviders(
 	container: Container,
 	authenticationService: IntegrationAuthenticationService,