+++ src/webviews/commitDetails/commitDetailsWebview.ts
@@ -755,7 +755,7 @@ export class CommitDetailsWebviewProvider
 		const {
 			pr: { url },
 		} = this.pullRequestContext;
-		return executeCommand<OpenPullRequestOnRemoteCommandArgs>(Commands.OpenPullRequestOnRemote, {
+		await executeCommand<OpenPullRequestOnRemoteCommandArgs>(Commands.OpenPullRequestOnRemote, {
 			pr: { url: url },
 			clipboard: clipboard,
 		});